reorder activerecord association and validation challenges
In the validations challenge we're asking the students to validate the
foreign_key, because they haven't yet created the belongs_to
associations. This is improper. When possible validations should be
placed on the association, here's an example:

````
 # incorrect
validates :dog_id, presence: true

 # correct
belongs_to :dog
validates :dog, presence: true

````

consider that the foreign_key is declared explicitly or derived from the
association declaration. AR is smart enough to validate the proper thing
when you tell it that the dog association is required.

So, in order to have students validate the CORRECT thing `dog` instead
of `dog_id`, we need to cover how to declare associations first.