move all gist content into the guide
teachers should be creating resources that live in the guide's resources
directory at this point. This is a one-time port of all the content that
was created previous to guides existing as repos.

By doing this we are creating the guide as the curation point for these
materials, this is important, because teachers leave, github usernames
change, and there's no reason that our content should disappear because
of those things.